requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.6.16
cffi==1.12.3
colorama==0.4.1
cryptography==2.7
decorator==4.4.0
defusedxml==0.6.0
emoji==0.5.3
entrypoints==0.3
future==0.17.1
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.14.1
Jinja2==2.10.1
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.0
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
pycparser==2.19
Pygments==2.4.2
pylint==2.3.1
pyrsistent==0.15.4
python-dateutil==2.8.0
python-telegram-bot==11.1.0
pywinpty==0.5.5
pyzmq==18.1.0
qtconsole==4.5.2
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
typed-ast==1.4.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
